// src/types/textField.ts
export type TextFieldTypes =
  | "text"
  | "password"
  | "email"
  | "number"
  | "tel"
  | "url"
  | "search"
  | "datetime-local"
  | "date"
  | "time"
  | "month"
  | "week";

export type InputModeType =
  | "search"
  | "tel"
  | "numeric"
  | "email"
  | "text"
  | "url"
  | "none"
  | "decimal"
  | undefined;